SnapdragonCamera: Use Glide for async bitmap loading
This is what Camera2 code uses.

CameraNext: Fix rotation not updated

The thumbnail generated by Glide is not updated
when user rotate the image manually.
This is because exif is rewrite using nio mmap,
however there is no nio unmap available in java.
Even though the file channel is closed, the rotation
modification is still in buffer, when Glide tries
to dtermine the orientation of the image, it still
reads the old data.

Use force to ensure data is sync to disk and buffer flushed.

HAM-1389

Fix memory leaks

	Two leaks were occuring. One was an unused Parameters
	in AndroidCameraManagerImpl that was holding onto excess
	PhotoModules. The second was Glide images for video previews that were never being cleared on recycle.

issue-id: CYNGNOS-2731

CameraNext: don't call Glide to load image after Camera exits

There is a race condition where the filmstrip can still be trying
to load a new image just as the camera exit, and Glide gets
upset if it is handed a destroyed activity context to work with.
So just check for that condition and go home.

Change-Id: I9e76ab40d23213c7a74c26779e345207d5fd7bae
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>